requirements.txt
{\rtf1\ansi\ansicpg949\cocoartf2580
\cocoatextscaling0\cocoaplatform0{\fonttbl\f0\fnil\fcharset0 HelveticaNeue;}
{\colortbl;\red255\green255\blue255;\red38\green38\blue38;}
{\*\expandedcolortbl;;\cssrgb\c20000\c20000\c20000;}
\paperw11900\paperh16840\margl1440\margr1440\vieww11520\viewh8400\viewkind0
\deftab720
\pard\pardeftab720\partightenfactor0

\f0\fs32 \cf2 \expnd0\expndtw0\kerning0
appnope==0.1.2\
asgiref==3.4.1\
backcall==0.2.0\
bcrypt==3.2.0\
certifi==2021.5.30\
cffi==1.14.6\
cryptography==35.0.0\
decorator==5.1.0\
Django==3.2.7\
django-cors-headers==3.9.0\
ipython==7.28.0\
jedi==0.18.0\
matplotlib-inline==0.1.3\
parso==0.8.2\
pexpect==4.8.0\
pickleshare==0.7.5\
prompt-toolkit==3.0.20\
ptyprocess==0.7.0\
pycparser==2.20\
Pygments==2.10.0\
PyJWT==2.2.0\
pytz==2021.3\
six==1.16.0\
sqlparse==0.4.2\
traitlets==5.1.0\
wcwidth==0.2.5}